#1029: runtime - consolidate gc
