+++ src/mu/gc/mod.rs
@@ -1,0 +1,12 @@
+//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
+//  SPDX-License-Identifier: MIT
+
+//! gc
+pub mod async_;
+pub mod cons;
+pub mod function;
+pub mod gc_;
+pub mod namespace;
+pub mod struct_;
+pub mod symbol;
+pub mod vector;
