+++ src/mu/types/function.rs
@@ -15,7 +15,6 @@ use {
             type_::Type,
         },
         namespaces::{
-            gc::{Gc as _, GcContext},
             heap::HeapRequest,
         },
         streams::writer::StreamWriter,
@@ -33,50 +32,6 @@ use {
 pub struct Function {
     pub arity: Tag, // number of required arguments
     pub form: Tag,  // list
-}
-
-pub trait Gc {
-    fn ref_form(_: &mut GcContext, _: Tag) -> Tag;
-    fn mark(_: &mut GcContext, _: &Env, _: Tag);
-    fn gc_ref_image(_: &mut GcContext, _: Tag) -> Self;
-}
-
-impl Gc for Function {
-    fn ref_form(context: &mut GcContext, func: Tag) -> Tag {
-        Self::gc_ref_image(context, func).form
-    }
-
-    fn mark(context: &mut GcContext, env: &Env, function: Tag) {
-        let mark = context.mark_image(function).unwrap();
-
-        if !mark {
-            let form = Self::ref_form(context, function);
-
-            context.mark(env, form);
-        }
-    }
-
-    fn gc_ref_image(context: &mut GcContext, tag: Tag) -> Self {
-        assert_eq!(tag.type_of(), Type::Function);
-
-        let heap_ref = &context.heap_ref;
-
-        match tag {
-            Tag::Indirect(fn_) => Self::new(
-                Tag::from_slice(
-                    heap_ref
-                        .image_slice(usize::try_from(fn_.image_id()).unwrap())
-                        .unwrap(),
-                ),
-                Tag::from_slice(
-                    heap_ref
-                        .image_slice(usize::try_from(fn_.image_id()).unwrap() + 1)
-                        .unwrap(),
-                ),
-            ),
-            Tag::Direct(_) => panic!(),
-        }
-    }
 }
 
 impl Function {
